Move Spark diversity metrics inside spark_evaluation .py; fix novelty; rewrite item novelty without cross join.
